多分Fでとれるかな？学校で試す

+++ Item.h
@@ -54,6 +54,6 @@ public:
 		float num = a.x - b.x;
 		float num2 = a.y - b.y;
 		float num3 = a.z - b.z;
-		return (float)std::sqrt(num * num + num2 * num2 + num3 * num3);
+		return std::sqrt(num * num + num2 * num2 + num3 * num3);
 	}
 };
